moved converter_mapping and get_converter into converter.py
