+++ src/arc/convert/__init__.py
@@ -1,4 +1,3 @@
-from typing import Dict, Type, Optional, Union
 from arc.errors import ConversionError
 from arc.convert.base_converter import BaseConverter
 from arc.convert.converters import *
@@ -14,20 +13,3 @@ __all__ = [
     "ListConverter",
     "FileConverter",
 ]
-
-converter_mapping: Dict[str, Type[BaseConverter]] = {
-    "str": StringConverter,
-    "int": IntConverter,
-    "float": FloatConverter,
-    "bytes": BytesConverter,
-    "bool": BoolConverter,
-    "sbool": StringBoolConverter,
-    "ibool": IntBoolConverter,
-    "list": ListConverter,
-}
-
-
-def get_converter(key: Union[str, type]) -> Optional[Type[BaseConverter]]:
-    if isinstance(key, type):
-        key = key.__name__
-    return converter_mapping.get(key)
